requirements.txt
numpy
fastapi
uvicorn[standard]
pdfminer.six
